added home folder and login page structure
